refactor: 함수 -> 화살표 함수로 구조 변경
