+++ src/components/AuthHelper/AuthHelper.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { AuthHelperContainer, Divider, StyledLink } from './AuthHelper.styled';
 
-export default function AuthHelper() {
+const AuthHelper = () => {
   return (
     <AuthHelperContainer>
       <StyledLink to="/signup">회원가입</StyledLink>
@@ -9,4 +9,6 @@ export default function AuthHelper() {
       <StyledLink to="/memberfind">아이디/비밀번호 찾기</StyledLink>
     </AuthHelperContainer>
   );
-}
+};
+
+export default AuthHelper;
