+++ src/pages/AllRecipe/AllRecipe.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { AiOutlinePlus } from 'react-icons/ai';
 import { getAllRecipe } from 'apis/request/recipe';
